bitmap,alter data type

+++ algorithms/bitmap_benchmark_test.go
@@ -3,8 +3,8 @@ package algorithms
 import "testing"
 
 func BenchmarkBitmap1(b *testing.B) {
-	max := int64(10_0000)
-	nums := []int64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000, 10_000, max}
+	max := int(10_0000)
+	nums := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000, 10_000, max}
 	bitmap := NewBitmap(max)
 	for _, v := range nums {
 		bitmap.Set(v)
@@ -19,8 +19,8 @@ func BenchmarkBitmap1(b *testing.B) {
 }
 
 func BenchmarkBitmap2(b *testing.B) {
-	max := int64(10_0000_0000)
-	nums := []int64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000, 10_000, max}
+	max := (10_0000_0000)
+	nums := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000, 10_000, max}
 	bitmap := NewBitmap(max)
 	for _, v := range nums {
 		bitmap.Set(v)
@@ -35,8 +35,8 @@ func BenchmarkBitmap2(b *testing.B) {
 }
 
 func BenchmarkBitmap3(b *testing.B) {
-	max := int64(50_0000_0000)
-	nums := []int64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000, 10_000, max}
+	max := int(50_0000_0000)
+	nums := []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 100, 1000, 10_000, max}
 	bitmap := NewBitmap(max)
 	for _, v := range nums {
 		bitmap.Set(v)
